requirements.txt
absl-py==2.1.0
asttokens==3.0.0
astunparse==1.6.3
Brotli==1.0.9
cachetools==5.5.0
certifi==2025.4.26
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
contourpy==1.2.0
cycler==0.11.0
debugpy==1.8.11
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.1.0
flatbuffers==24.12.23
fonttools==4.51.0
gast==0.4.0
google-auth==2.37.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.69.0
h5py==3.12.1
idna==3.10
imageio==2.33.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
importlib_resources==6.4.0
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==18.1.1
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
moviepy==1.0.3
nest_asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==24.2
parso==0.8.4
pickleshare==0.7.5
pillow==11.0.0
pip==24.3.1
platformdirs==4.3.6
ply==3.11
proglog==0.1.10
prompt_toolkit==3.0.48
protobuf==3.19.6
psutil==6.1.1
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.19.1
pyparsing==3.2.0
PyQt5==5.15.10
PyQt5-sip==12.13.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pywin32==307
pyzmq==26.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.6.1
scipy==1.13.1
setuptools==75.1.0
sip==6.7.12
six==1.17.0
stack_data==0.6.3
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
threadpoolctl==3.5.0
tomli==2.0.1
tornado==6.4.2
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
unicodedata2==15.1.0
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.44.0
win-inet-pton==1.1.0
wrapt==1.17.2
zipp==3.21.0
